Move scan_cached_blocks out of sqlite crate.
