+++ zcash_client_backend/Cargo.toml
@@ -33,6 +33,8 @@ protobuf-codegen-pure = "2.15"
 [dev-dependencies]
 rand_core = "0.5.1"
 rand_xorshift = "0.2"
+tempfile = "3.1.0"
+zcash_client_sqlite = { version = "0.2", path = "../zcash_client_sqlite" }
 
 [features]
 test-dependencies = ["proptest", "zcash_primitives/test-dependencies"]
